Fix ValueError in test_export_from_dgm_bucket resident-ratio check
vb_active_perc_mem_resident returns a decimal string, but the wait
loop passed it to int() instead of float(), raising
"ValueError: invalid literal for int() with base 10" whenever the
resident ratio wasn't a whole number.

Reviewed-on: https://review.couchbase.org/c/testrunner/+/248254
Tested-by: <anisha.sinha@couchbase.com>

+++ pytests/clitest/importexporttest.py
@@ -1014,7 +1014,7 @@ class ImportExportTests(CliBaseTest):
                     if self.dgm_run:
                         res_status = stats_all_buckets[bucket.name].get_stats([self.master],
                                      bucket, '', 'vb_active_perc_mem_resident')[self.master]
-                        while int(res_status) > self.active_resident_threshold:
+                        while float(res_status) > self.active_resident_threshold:
                             self.sleep(5)
                             res_status = stats_all_buckets[bucket.name].get_stats([self.master],
                                          bucket, '',
